src/requirements.txt
numpy==1.24.3
requests==2.30.0
edge-tts==6.1.9
pygame==2.5.2
pyinstaller==6.3.0
vosk==0.3.45
sounddevice==0.4.6
aiohttp==3.9.1
Pillow==10.1.0 